Refactor mapView
